Change the definition of doub to long double to get higher precesion.

+++ comb.h
@@ -17,7 +17,7 @@
 #include "atom.cpp" //for atom struct
 using namespace gmm;
 using namespace std;
-typedef double doub;
+typedef long double doub;
 extern int sweep(int steps,int total_steps,doub PeakPower,doub convergence,doub convergence_threshold,int conS,int expN,int n1, int n2,int Msteps,doub detune);
 const doub pi=3.14159265358979323846264338327950288419716939937511;
 #endif // COMB_H_INCLUDED
